// frontend/src/components/ui/button.d.ts
import * as React from 'react';
export interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
  variant?: 'default'|'destructive'|'outline'|'secondary'|'ghost'|'link';
  size?: 'default'|'sm'|'lg'|'icon';
  asChild?: boolean;
}
export const Button: React.ForwardRefExoticComponent<ButtonProps & React.RefAttributes<HTMLButtonElement>>;
export const buttonVariants: (props?: Pick<ButtonProps,'variant'|'size'|'className'>) => string;